% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{write.docx}
\alias{write.docx}
\title{write data.frame to docx table}
\usage{
write.docx(x, file, layout = "portrait", overwrite = TRUE, colour = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{file}{file to save to}

\item{layout}{"portrait" or "landscape" page orientation}

\item{overwrite}{logical. Overwrite existing file.}

\item{colour}{A valid hex colour code starting with '#' eg. '#4f81BD'}
}
\value{
Workbook object
}
\description{
write data.frame to docx table
}
\examples{
write.docx(x = iris, "testing.docx", layout = "portrait", colour = NULL)
write.docx(x = iris, "testing1.docx", layout = "portrait", colour = "#4f81BD")
write.docx(x = iris, "testing2.docx", layout = "landscape", colour = "#C0504D")
}
\author{
Alexander Walker
}

